backend/pyproject.toml
[project]
name = "django-project-template"
version = "1.2"
description = "A modern, production-ready Django project template with REST API, Redis caching, Prometheus metrics, AWS S3 integration, and DRF support."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.39.13",
    "cryptography>=41.0.0",
    "django>=5.2.7",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-import-export>=4.3.9",
    "django-jazzmin>=3.0.1",
    "django-prometheus>=2.4.1",
    "django-redis>=6.0.0",
    "django-storages>=1.14.6",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "drf-standardized-errors>=0.15.0",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "prometheus-client>=0.22.1",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "django-debug-toolbar>=5.2.0",
    "flake8>=7.3.0",
]
